requirements.txt
streamlit==1.17.0
Pillow==9.4.0
requests==2.28.2
botocore
boto3
python-dotenv
